refactor: move modal components in accordance with atomic design
